Favorites and Blacklist implemented, styles upgraded, code optimizations done
