+++ Boozer (Android Studio)/.idea/dictionaries/rafaa.xml
@@ -1,0 +1,3 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="rafaa" />
+</component>
