--- Boozer (Android Studio)/README.md
@@ -1,50 +1,0 @@
-# _BoozeR_
-## ¿Alguna vez has tenido problemas a la hora de elegir qué beber?
-### Quizás no quieras gastarte mucho dinero. A lo mejor quieres probar un ron nuevo. O quizás simplemente no quieras algo muy cargado.
-Para esos casos, y para muchos más, nació ***BoozeR***.
-
-## _¿Qué es BoozeR?_
-BoozeR es una aplicación móvil desarrollada en Android Studio que busca facilitar la vida a los usuarios que quieran descubrir bebidas nuevas.
-
-Principalmente, la aplicación está destinada a gestionar bebidas alcohólicas. No obstante, también será posible gestionar bebidas sin alcohol, así como complementos para las mismas. </br>
-
-## _Secciones de BoozeR_
-- Índice</br></br>
-<img src="media/index.jpg" height="600px">
-</br>
-
-- Registro</br></br>
-<img src="media/register.jpg" height="600px">
-</br>
-
-- Inicio de Sesión</br></br>
-<img src="media/login.jpg" height="600px">
-</br>
-
-- Vista Principal, que a su vez contendrá varios fragmentos:</br>
-  - Catálogo de Bebidas</br></br>
-  <img src="media/catalogue.jpg" height="600px">
-  </br></br>
-  
-  - Calculador de Alcohol en Sangre</br></br>
-  <img src="media/calculator.jpg" height="600px">
-  </br>
-  
-- Perfil</br></br>
-<img src="media/profile.jpg" height="600px">
-</br></br>
-
-  - Editar Cuenta
-<img src="media/profile_edit.jpg" height="600px">
-</br></br>
-
-  - Borrar Cuenta
-<img src="media/profile_delete.jpg" height="600px">
-</br>
-
-## _TODO_
-En próximas actualizaciones, se añadira un nuevo fragmento llamado _coctelería_, en el que podrás crear tus propios cócteles a partir de bebidas ya existentes.
-
-## _Contacto_
-En caso de no entender algo, o si sencillamente quieres dar ideas para mejorar la aplicación, no dudes en contactarme a trav-es de mi correo, rafa.ar2000@gmail.com.
-
