+++ Boozer (Android Studio)/app/src/main/res/layout/activity_drink.xml
@@ -13,7 +13,7 @@
     android:layout_height="wrap_content"
     android:background="@color/darkBackground"
     android:fillViewport="true"
-    tools:context=".Actividades.DrinkActivity">
+    tools:context=".Activities.DrinkActivity">
     <!--Loading gif-->
     <ImageView
         android:id="@+id/loadingDrink"
@@ -90,12 +90,40 @@
         android:id="@+id/drinkDetails"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:paddingTop="20dp"
+        android:paddingTop="10dp"
         android:textColor="@color/white"
         android:textSize="40sp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/drinkPrice" />
+
+    <!-- Toggle favorite button -->
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btnToggleFav"
+        style="@style/Widget.MaterialComponents.Button.TextButton"
+        android:layout_width="50dp"
+        android:layout_height="62dp"
+        android:layout_marginTop="10dp"
+        android:drawableTop="@drawable/like"
+        android:drawableTint="@color/disabled"
+        android:visibility="invisible"
+        app:layout_constraintEnd_toStartOf="@+id/btnToggleDis"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/drinkDetails" />
+
+    <!-- Toggle dislike button -->
+    <com.google.android.material.button.MaterialButton
+        android:id="@+id/btnToggleDis"
+        style="@style/Widget.MaterialComponents.Button.TextButton"
+        android:layout_width="50dp"
+        android:layout_height="62dp"
+        android:layout_marginTop="10dp"
+        android:drawableTop="@drawable/dislike"
+        android:alpha=".25"
+        android:visibility="invisible"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toEndOf="@+id/btnToggleFav"
+        app:layout_constraintTop_toBottomOf="@+id/drinkDetails" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
