+++ Boozer (Android Studio)/app/src/main/res/layout/activity_index.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/darkBackground"
-    tools:context=".Actividades.IndexActivity">
+    tools:context=".Activities.IndexActivity">
 
     <!-- Logo -->
     <ImageView
