+++ Boozer (Android Studio)/app/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/darkBackground"
-    tools:context=".Actividades.LoginActivity">
+    tools:context=".Activities.LoginActivity">
 
     <!-- campo EMAIL -->
     <com.google.android.material.textfield.TextInputLayout
