+++ Boozer (Android Studio)/app/src/main/res/layout/fragment_catalogue.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fragmentos.CatalogueFragment"
+    tools:context=".Fragments.CatalogueFragment"
     android:scrollbarThumbVertical="@color/colorPrimaryDark">
 
     <!--Loading gif-->
