+++ Boozer (Android Studio)/app/src/main/res/layout/item_layout.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="325dp">
+    android:layout_height="335dp">
 
     <com.google.android.material.card.MaterialCardView
         android:layout_width="match_parent"
@@ -21,11 +21,11 @@
         <TextView
             android:id="@+id/itemName"
             android:layout_width="match_parent"
-            android:layout_height="20dp"
-            android:layout_marginTop="240dp"
+            android:layout_height="28sp"
+            android:layout_marginTop="250dp"
             android:textStyle="bold"
             android:gravity="center_horizontal"
-            android:textSize="15sp" />
+            android:textSize="22sp" />
     </com.google.android.material.card.MaterialCardView>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
