+++ Boozer (Android Studio)/app/src/main/res/values/colors.xml
@@ -2,9 +2,12 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <color name="colorPrimary">#201f1e</color>
     <color name="colorPrimaryDark">#4E0000</color>
+    <color name="bottomNav">#6c1b1b</color>
     <color name="colorAccent">#A83F3F</color>
     <color name="darkBackground">#3F3F3F</color>
     <color name="white">#FFFFFF</color>
     <color name="black">#000000</color>
+    <color name="favorite">#FFB100</color>
+    <color name="disabled">#464646</color>
     <color name="mtrl_textinput_default_box_stroke_color" tools:ignore="PrivateResource">#201f1e</color>
 </resources>
