+++ Boozer (Android Studio)/app/src/main/res/values/styles.xml
@@ -6,7 +6,6 @@
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
-        <item name="android:navigationBarColor">@color/colorPrimary</item>
         <item name="android:textColor">@color/white</item>
         <item name="android:backgroundTint">@color/colorPrimary</item>
         <item name="boxStrokeColor">#fff</item>
